nits(infiniband): remove id package, make code ordering consistent
