--- components/accelerator/nvidia/infiniband/id/id.go
@@ -1,4 +1,0 @@
-// Package id provides the ID for the NVIDIA InfiniBand component.
-package id
-
-const Name = "accelerator-nvidia-infiniband"
